feat(homeTableManageColumns): Organized code for home table state management
