+++ src/Analysis/GWASResultsV2/Views/Home/Home.stories.jsx
@@ -5,7 +5,7 @@ import { rest } from 'msw';
 import Home from './Home';
 import PHASES from '../../Utils/PhasesEnumeration';
 import TableData from '../../TestData/TableData';
-import InitialHomeTableState from '../../Utils/InitialHomeTableState';
+import InitialHomeTableState from '../../Utils/StateManagement/InitialHomeTableState';
 
 const setCurrentView = (input) => {
   alert(`setCurrentView called with ${input}`);
